header version macro fix pointed out by lubos dolezel
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@36441 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSCalendar.m
@@ -79,8 +79,8 @@ typedef struct {
   NSString      *localeID;
   NSTimeZone    *tz; 
   void          *cal;
-  NSInteger     firstWeekday;
-  NSInteger     minimumDaysInFirstWeek;
+  int32_t       firstWeekday;
+  int32_t       minimumDaysInFirstWeek;
 } Calendar; 
 #define my ((Calendar*)_NSCalendarInternal)
 #define aac ((Calendar*)(autoupdatingCalendar->_NSCalendarInternal))
